# perf: remove need for channels in the sender recovery stage