This PR proposes a slightly different approach at the parallelization of the reading of the transactions: instead of using a single cursor for a big range, it slices the range and opens a cursor for each sub-range. This alternative approach makes the use of channels unnecessary, thus removing some overhead.

Unfortunately due to a separate error that seems to happen on my machine I wasn't able to run tests to validate nothing is broken nor the benchmarks in order to compare performance.