@mattsse 2023-07-05
what's the performance impact of this