@auterium 2023-07-05
> what's the performance impact of this

Unfortunately I couldn't measure it as I haven't been able to run benchmarks in my machine as running `cargo bench` only shows that all tests were ignored. Any pointers on how to run them locally would be much appreciated.

My expectation would be to reduce some overhead that's incurred from the use of the channels, probably some memory moves/clones would be avoided as well. Unsure if the use of multiple read cursors has a meaningful positive impact.