@joshieDo 2023-07-05
You can try running `hyperfine` over just the stage, if you're having issues with `cargo bench`. 

eg. 
```bash
hyperfine reth stage tx-lookup --from 1 --to 10000
```