@auterium 2023-07-05
@joshieDo didn't work:
```
$ ./target/release/reth stage tx-lookup --from 1 --to 10000
error: unrecognized subcommand 'tx-lookup'

Usage: reth stage [OPTIONS] <COMMAND>

For more information, try '--help'.
```

The error I get when I try to run the benches (same as when trying to run a test)
```
$ cargo bench --package reth-stages --bench criterion --features test-utils
    Finished bench [optimized] target(s) in 0.84s
     Running benches/criterion.rs (/home/user/reth/target/release/deps/criterion-bd88f246948d03c9)
thread 'main' panicked at 'Not able to create the database file.: Failed to open database: 37

Location:
    /home/user/reth/crates/storage/db/src/lib.rs:125:18', /home/user/reth/crates/storage/db/src/lib.rs:187:47
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
error: bench failed, to rerun pass `-p reth-stages --bench criterion`
```
I'm running this in a WSL2