@onbjerg 2023-07-05
@auterium The command is `reth stage run tx-lookup --from 1 --to 10000`. Note that it requires downloaded headers and bodies though (as this builds an index using another tx table)

In terms of the other error, is there a chance you could reach out to me on Telegram at @onbjerg? We have a similar open issue for WSL, but I am not entirely sure how to debug it. Would be great if we could jam a bit together on it? I assume it's either a permission error or a pathing error :)