@onbjerg 2023-07-05
Ran the benchmarks on my machine.

Take away: The performance is slightly worse, or around the same. It's somewhat hard to tell because the variance on these benchmarks is pretty high for whatever reason. Running the exact same branch (even on main) sometimes yields +-30% difference.

It might be because we collect the results in a `Vec` just to unfurl it after. Afaik rayon has a method to parallelize the work + send the results over a channel, so this might be an idea. We would not allocate a lot of channels, just one. This is just conjecture though

**main**

Run 1

```
Stages/SendersRecovery-batch-1000
                        time:   [37.534 ms 38.799 ms 40.285 ms]
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) high mild
  1 (10.00%) high severe
Stages/SendersRecovery-batch-10000
                        time:   [49.665 ms 50.527 ms 51.285 ms]
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low severe
Stages/SendersRecovery-batch-100000
                        time:   [51.090 ms 52.780 ms 54.757 ms]
Stages/SendersRecovery-batch-250000
                        time:   [50.149 ms 51.855 ms 53.591 ms]
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
```

Run 2

```
Stages/SendersRecovery-batch-1000
                        time:   [30.965 ms 31.282 ms 31.616 ms]
                        change: [-61.393% -51.759% -34.872%] (p = 0.00 < 0.05)
                        Performance has improved.
Stages/SendersRecovery-batch-10000
                        time:   [30.643 ms 30.845 ms 31.141 ms]
                        change: [-9.1515% -6.9501% -4.1583%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
Stages/SendersRecovery-batch-100000
                        time:   [30.559 ms 31.218 ms 31.588 ms]
                        change: [-11.122% -9.1172% -7.1083%] (p = 0.00 < 0.05)
                        Performance has improved.
Stages/SendersRecovery-batch-250000
                        time:   [30.627 ms 31.084 ms 31.472 ms]
                        change: [-11.601% -9.8231% -7.9093%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Run 3

```
Stages/SendersRecovery-batch-1000
                        time:   [29.957 ms 30.412 ms 30.839 ms]
                        change: [-4.6193% -3.1516% -1.7266%] (p = 0.00 < 0.05)
                        Performance has improved.
Stages/SendersRecovery-batch-10000
                        time:   [30.556 ms 30.745 ms 30.890 ms]
                        change: [-5.7978% -2.7678% -0.4773%] (p = 0.06 > 0.05)
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low mild
Stages/SendersRecovery-batch-100000
                        time:   [31.117 ms 31.448 ms 31.991 ms]
                        change: [+0.3564% +2.3855% +4.6315%] (p = 0.05 < 0.05)
                        Change within noise threshold.
Stages/SendersRecovery-batch-250000
                        time:   [30.769 ms 30.923 ms 31.226 ms]
                        change: [-0.4514% +1.6974% +4.5647%] (p = 0.21 > 0.05)
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
```

Run 4

```
Stages/SendersRecovery-batch-1000
                        time:   [32.257 ms 32.526 ms 32.982 ms]
                        change: [+5.7752% +8.1309% +10.667%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Stages/SendersRecovery-batch-10000
                        time:   [30.997 ms 31.663 ms 32.697 ms]
                        change: [+1.1045% +3.4276% +5.7494%] (p = 0.01 < 0.05)
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Stages/SendersRecovery-batch-100000
                        time:   [31.044 ms 31.431 ms 31.989 ms]
                        change: [-1.4225% +0.5492% +2.8081%] (p = 0.64 > 0.05)
                        No change in performance detected.
Stages/SendersRecovery-batch-250000
                        time:   [29.411 ms 29.767 ms 30.630 ms]
                        change: [-4.9525% -1.8322% +1.2555%] (p = 0.31 > 0.05)
                        No change in performance detected.
```

Run 5

```
Stages/SendersRecovery-batch-1000
                        time:   [30.379 ms 30.816 ms 31.320 ms]
                        change: [-7.8451% -5.6067% -3.5526%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 10 measurements (30.00%)
  1 (10.00%) low mild
  2 (20.00%) high mild
Stages/SendersRecovery-batch-10000
                        time:   [29.539 ms 29.731 ms 30.081 ms]
                        change: [-7.9615% -5.6632% -3.4618%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Stages/SendersRecovery-batch-100000
                        time:   [30.088 ms 30.334 ms 30.573 ms]
                        change: [-5.4673% -1.7100% +3.5125%] (p = 0.57 > 0.05)
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
Stages/SendersRecovery-batch-250000
                        time:   [30.830 ms 31.210 ms 31.737 ms]
                        change: [-1.2487% +1.8134% +4.7693%] (p = 0.28 > 0.05)
                        No change in performance detected.
```

**This PR**

Run 1

```
Stages/SendersRecovery-batch-1000
                        time:   [32.929 ms 33.245 ms 33.499 ms]
                        change: [+5.3412% +6.9373% +8.5605%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Stages/SendersRecovery-batch-10000
                        time:   [33.202 ms 33.344 ms 33.470 ms]
                        change: [+10.089% +11.852% +13.448%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low mild
Stages/SendersRecovery-batch-100000
                        time:   [32.945 ms 33.218 ms 33.459 ms]
                        change: [+1.1740% +6.5664% +10.160%] (p = 0.02 < 0.05)
                        Performance has regressed.
Stages/SendersRecovery-batch-250000
                        time:   [29.735 ms 30.955 ms 33.245 ms]
                        change: [+0.1833% +4.2606% +8.8142%] (p = 0.09 > 0.05)
                        No change in performance detected.
```

Run 2

```
Stages/SendersRecovery-batch-1000
                        time:   [32.677 ms 33.110 ms 33.455 ms]
                        change: [-1.4213% -0.1893% +0.9812%] (p = 0.77 > 0.05)
                        No change in performance detected.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) low mild
  1 (10.00%) high mild
Stages/SendersRecovery-batch-10000
                        time:   [28.451 ms 28.606 ms 28.867 ms]
                        change: [-14.830% -13.826% -12.850%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low mild
Stages/SendersRecovery-batch-100000
                        time:   [32.971 ms 33.296 ms 33.733 ms]
                        change: [-4.2021% -1.4313% +1.2871%] (p = 0.40 > 0.05)
                        No change in performance detected.
Found 3 outliers among 10 measurements (30.00%)
  2 (20.00%) low severe
  1 (10.00%) high mild
Stages/SendersRecovery-batch-250000
                        time:   [32.990 ms 33.124 ms 33.280 ms]
                        change: [-2.3208% +1.6661% +6.1809%] (p = 0.49 > 0.05)
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low mild
```

Run 3

```
Stages/SendersRecovery-batch-1000
                        time:   [36.402 ms 39.034 ms 41.081 ms]
                        change: [+16.208% +20.750% +23.735%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low severe
Stages/SendersRecovery-batch-10000
                        time:   [32.923 ms 33.191 ms 33.450 ms]
                        change: [+14.784% +15.903% +17.166%] (p = 0.00 < 0.05)
                        Performance has regressed.
Stages/SendersRecovery-batch-100000
                        time:   [33.214 ms 33.655 ms 33.916 ms]
                        change: [-0.0134% +2.6907% +6.1052%] (p = 0.12 > 0.05)
                        No change in performance detected.
Stages/SendersRecovery-batch-250000
                        time:   [33.107 ms 33.627 ms 34.119 ms]
                        change: [+0.1104% +1.9470% +3.8806%] (p = 0.07 > 0.05)
                        No change in performance detected.
```

Run 4

```
Stages/SendersRecovery-batch-1000
                        time:   [35.329 ms 35.804 ms 36.328 ms]
                        change: [-13.053% -10.487% -6.9558%] (p = 0.00 < 0.05)
                        Performance has improved.
Stages/SendersRecovery-batch-10000
                        time:   [33.909 ms 34.098 ms 34.507 ms]
                        change: [+2.6052% +3.9043% +5.2993%] (p = 0.00 < 0.05)
                        Performance has regressed.
Stages/SendersRecovery-batch-100000
                        time:   [33.276 ms 33.888 ms 34.816 ms]
                        change: [-0.4712% +1.5942% +3.8985%] (p = 0.17 > 0.05)
                        No change in performance detected.
Stages/SendersRecovery-batch-250000
                        time:   [33.917 ms 34.194 ms 34.494 ms]
                        change: [+0.1753% +2.2784% +4.5691%] (p = 0.07 > 0.05)
                        No change in performance detected.
```

Run 5

```
Stages/SendersRecovery-batch-1000
                        time:   [33.418 ms 34.083 ms 34.534 ms]
                        change: [-8.6696% -6.5266% -4.5708%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low mild
Stages/SendersRecovery-batch-10000
                        time:   [34.921 ms 35.748 ms 36.291 ms]
                        change: [-1.0110% +1.2769% +3.8423%] (p = 0.35 > 0.05)
                        No change in performance detected.
Stages/SendersRecovery-batch-100000
                        time:   [33.248 ms 33.641 ms 34.322 ms]
                        change: [-2.0492% +0.4312% +2.9552%] (p = 0.75 > 0.05)
                        No change in performance detected.
Stages/SendersRecovery-batch-250000
                        time:   [33.160 ms 33.553 ms 33.874 ms]
                        change: [-4.9839% -3.2617% -1.6826%] (p = 0.00 < 0.05)
                        Performance has improved.
```